+++ pkg/tempo/config_test.go
@@ -447,7 +447,7 @@ spanmetrics:
     - name: http.method
       default: GET
     - name: http.status_code
-  prom_instance: tempo
+  metrics_instance: tempo
 `,
 			expectedConfig: `
 receivers:
@@ -463,7 +463,7 @@ exporters:
       max_elapsed_time: 60s
   remote_write:
     namespace: tempo_spanmetrics
-    prom_instance: tempo
+    metrics_instance: tempo
 processors:
   spanmetrics:
     metrics_exporter: remote_write
@@ -535,7 +535,7 @@ remote_write:
   - endpoint: example.com:12345
 spanmetrics:
   handler_endpoint: "0.0.0.0:8889"
-  prom_instance: tempo
+  metrics_instance: tempo
 `,
 			expectedError: true,
 		},
@@ -858,7 +858,7 @@ spanmetrics:
     - name: http.method
       default: GET
     - name: http.status_code
-  prom_instance: tempo
+  metrics_instance: tempo
 automatic_logging:
   spans: true
 batch:
@@ -906,7 +906,7 @@ spanmetrics:
     - name: http.method
       default: GET
     - name: http.status_code
-  prom_instance: tempo
+  metrics_instance: tempo
 automatic_logging:
   spans: true
 batch:
@@ -963,7 +963,7 @@ spanmetrics:
     - name: http.method
       default: GET
     - name: http.status_code
-  prom_instance: tempo
+  metrics_instance: tempo
 automatic_logging:
   spans: true
 batch: