+++ CHANGELOG.md
@@ -66,6 +66,9 @@ for specific instructions.
   hyphen in the name to be consistent with how Kubernetes refers to resources.
   (@rfratto)
 
+- [CHANGE] Breaking change: `prom_instance` in the spanmetrics config is now
+  named `metrics_instance`. (@rfratto)
+
 - [DEPRECATION] The `loki` key at the root of the config file has been
   deprecated in favor of `logs`. `loki`-named fields in `automatic_logging`
   have been renamed accordinly: `loki_name` is now `logs_instance_name`,